backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.1
httpx==0.25.2
python-dateutil==2.8.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
cryptography==41.0.7

